+++ src/optimize.rs
@@ -128,6 +128,7 @@ fn call_fn_with_constant_arguments(
             (hash_fn, 0),
             arg_values.iter_mut().collect::<StaticVec<_>>().as_mut(),
             false,
+            false,
             None,
             0,
         )